Replace httplib with Boost.Beast HTTP module; add Win11 SDK target; suppress 3rd-party warnings
- cmake/deps.cmake: add Boost 1.91.0 via CPM (Beast+ASIO); mark all CPM
  package includes as INTERFACE_SYSTEM_INCLUDE_DIRECTORIES to suppress
  -Wsign-conversion and other diagnostics on third-party headers
- cmake/compiler_flags.cmake: add _WIN32_WINNT=0x0A00 / WINVER=0x0A00
  globally on WIN32 to target Windows 10/11 and silence ASIO fallback noise
- include/sub0llm/http/client.hpp: new Beast-backed sync HTTP client
  (Post, set_connection_timeout, set_read_timeout) replacing httplib::Client
- include/sub0llm/http/server.hpp: new Beast-backed sync HTTP server
  (Get/Post routes, set_pre_routing_handler, listen) replacing httplib::Server;
  OPTIONS preflight handled automatically via pre-routing headers
- chapters/ch24_real_training/main.cpp + CMakeLists.txt: migrate from
  httplib::Client to sub0llm::http::Client; link Boost::beast instead of httplib
- tools/server/main.cpp + tools/CMakeLists.txt: migrate from httplib::Server
  to sub0llm::http::Server; remove new_task_queue; threads passed to listen()
- diffusion/chapters/ch32_tree_predictor/viz_server.cpp + CMakeLists.txt:
  full Beast refactor (already present); link Boost::beast Boost::system

+++ tools/cli/main.cpp
@@ -59,7 +59,7 @@ struct CliConfig {
     bool        greedy      = false;
     bool        interactive = false;
     bool        f32         = false;  // --f32: load full f32 weights (default: Q8 int8)
-    uint64_t    seed        = 42;
+    uint32_t    seed        = 42;
 };
 
 struct ModelArch {
